ambulance-notifier-service/requirements.txt
bcrypt==3.2.0
bidict==0.21.2
cffi==1.14.5
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
pycparser==2.20
PyMySQL==1.0.2
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
simple-websocket==0.2.0
six==1.16.0
SQLAlchemy==1.4.17
Werkzeug==2.0.1
wsproto==1.0.0
opencv-python==4.5.2.52
tensorflow==2.5.0
pytube==10.8.3
scikit-image==0.18.1
pillow==8.2.0